Move to a builder pattern for creating a BKTree.
