+++ src/Workspaces/Core/Portable/Workspaces.csproj
@@ -949,6 +949,7 @@
       <SubType>
       </SubType>
     </None>
+    <Compile Include="Utilities\BKTree.Builder.cs" />
     <PublicAPI Include="PublicAPI.Shipped.txt" />
     <PublicAPI Include="PublicAPI.Unshipped.txt" />
   </ItemGroup>
